meta-nxp3220: adds macro to build swupdate image with bbappend files
This patch adds macro to build swupdate image with bbappend files

- SWU_DAEMON_SERVICES_REPLACE: replace meta-swupdate's service daemon.
- SWU_SUPPORT_MAKE_SWU_IMAGE: build swu image with 'sw_image.sh' script

Reviewed-on: https://review.gerrithub.io/c/NexellCorp/yocto_meta-nexell/+/478145

+++ meta-nxp3220/recipes-bsp/images/nexell-recoveryimg.bb
@@ -33,7 +33,7 @@ DEPENDS_append = " update-rc.d-native"
 IMAGE_INSTALL_append = " \
 	swupdate \
 	swupdate-tools \
-	swupdate-recovery \
+	${@ 'swupdate-recovery' if d.getVar('SWU_DAEMON_SERVICES_REPLACE') else ''} \
 	udev \
 	util-linux-agetty \
 	"
@@ -71,19 +71,21 @@ postprocess_recovery_image () {
 			>> ${rootdir}${sysconfdir}/fstab
 	fi
 
-	# remove not necessary services
-	#
-	update-rc.d -r ${rootdir} -f psplash.sh remove
+	if [ ! -z "${SWU_DAEMON_SERVICES_REPLACE}" ] && [ "${SWU_DAEMON_SERVICES_REPLACE}" = "true" ]; then
+		# remove not necessary services
+		#
+		update-rc.d -r ${rootdir} -f psplash.sh remove
 
-	# remove not necessary udev rules
-	#
-	rm ${rootdir}/lib/udev/rules.d/*
+		# remove not necessary udev rules
+		#
+		rm ${rootdir}/lib/udev/rules.d/*
 
-	if [ -f ${rootdir}${sysconfdir}/udev/rules.d/local.rules ]; then
-		rm ${rootdir}${sysconfdir}/udev/rules.d/local.rules
-	fi
-	if [ -f ${rootdir}${sysconfdir}/udev/rules.d/touchscreen.rules ]; then
-		rm ${rootdir}${sysconfdir}/udev/rules.d/touchscreen.rules
+		if [ -f ${rootdir}${sysconfdir}/udev/rules.d/local.rules ]; then
+			rm ${rootdir}${sysconfdir}/udev/rules.d/local.rules
+		fi
+		if [ -f ${rootdir}${sysconfdir}/udev/rules.d/touchscreen.rules ]; then
+			rm ${rootdir}${sysconfdir}/udev/rules.d/touchscreen.rules
+		fi
 	fi
 }
